created folder for tests and implemented 3 tests to run in node.js
