+++ OLDbotTests.js
@@ -1,0 +1,81 @@
+// //  Verify Discord.js integration by creating a test function that creates a new Discord client
+// // Import necessary modules
+// const { Client, IntentsBitField } = require('discord.js');
+
+// // Custom assertion function to compare actual and expected values
+// function assertEqual(actual, expected, message) {
+//   if (actual !== expected) {
+//     throw new Error(`${message}: Expected ${expected}, but got ${actual}`);
+//   }
+// }
+
+// // Test function to create a new Discord client
+// function testDiscordIntegration() {
+//   try {
+//     console.log('Initializing Discord client...');
+
+//     // Arrange: Define bot intents
+//     console.log('Defining bot intents...');
+//     const botIntents = new IntentsBitField();
+//     botIntents.add(
+//       IntentsBitField.Flags.Guilds,
+//       IntentsBitField.Flags.GuildMessages,
+//       IntentsBitField.Flags.GuildMessageTyping,
+//       IntentsBitField.Flags.GuildEmojisAndStickers,
+//       IntentsBitField.Flags.MessageContent,
+//       IntentsBitField.Flags.GuildMembers,
+//       IntentsBitField.Flags.DirectMessages,
+//       IntentsBitField.Flags.DirectMessageReactions,
+//       IntentsBitField.Flags.DirectMessageTyping,
+//     );
+
+//     // Act: Try to create a new Discord client
+//     console.log('Creating Discord client...');
+//     const client = new Client({ intents: botIntents });
+
+//     // Assert: Check if client is an instance of Client
+//     console.log('Asserting Discord client creation...');
+//     assertEqual(typeof client, 'object', 'Client type');
+//     assertEqual(client instanceof Client, true, 'Client instance');
+
+//     console.log('Discord.js integration test passed successfully.');
+//   } catch (error) {
+//     console.error('Discord.js integration test failed:', error.message);
+//   }
+// }
+
+// // Call the test function
+// testDiscordIntegration();
+
+
+// // ---------------------------------------------------------------------------------
+// // // Ensure that the OpenAI library is correctly integrated by creating a test function that attempts to use the OpenAI API to create a simple chat completion or query.
+
+// // // OpenAI Library Integration Test
+// // // Import necessary modules
+// // const { OpenAI } = require('openai');
+
+// // // Test function to check OpenAI library integration
+// // function testOpenAIIntegration() {
+// //   try {
+// //     console.log('Initializing OpenAI integration test...');
+
+// //     // Act: Try to create an instance of OpenAI
+// //     const openai = new OpenAI({ apiKey: 'OpenAI_API_Key_Here' });
+
+// //     // Assert: Check if openai is an instance of OpenAI
+// //     console.log('Asserting OpenAI integration...');
+// //     assertEqual(typeof openai, 'object', 'OpenAI type');
+// //     assertEqual(openai instanceof OpenAI, true, 'OpenAI instance');
+
+// //     console.log('OpenAI library integration test passed successfully.');
+// //   } catch (error) {
+// //     console.error('OpenAI library integration test failed:', error.message);
+// //   }
+// // }
+
+// // // Call the test function
+// // testOpenAIIntegration();
+
+// // // --------------------------------------------------------------------------
+
